bug fix for predictive date in case of monthly active full
In case only one yearly monthly was selected, the predicted end date
might be too short. Optimized by calculating the max dist between
months, and configuring the predictive date in relation with that. In
most cases, this will actually shorten calculation time but in rare
configuration this will put the predict date almost twice as far